plugin: cleanup `proposed` api files
