+++ packages/plugin/src/theia.proposed.findTextInFiles.d.ts
@@ -1,0 +1,178 @@
+// *****************************************************************************
+// Copyright (C) 2023 Ericsson and others.
+//
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License v. 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0.
+//
+// This Source Code may also be made available under the following Secondary
+// Licenses when the conditions for such availability set forth in the Eclipse
+// Public License v. 2.0 are satisfied: GNU General Public License, version 2
+// with the GNU Classpath Exception which is available at
+// https://www.gnu.org/software/classpath/license.html.
+//
+// SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0
+// *****************************************************************************
+
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+// code copied and modified from https://github.com/microsoft/vscode/blob/1.77.0/src/vscode-dts/vscode.proposed.findTextInFiles.d.ts
+
+export module '@theia/plugin' {
+
+    /**
+     * Options that can be set on a findTextInFiles search.
+     */
+    export interface FindTextInFilesOptions {
+        /**
+         * A [glob pattern](#GlobPattern) that defines the files to search for. The glob pattern
+         * will be matched against the file paths of files relative to their workspace. Use a [relative pattern](#RelativePattern)
+         * to restrict the search results to a [workspace folder](#WorkspaceFolder).
+         */
+        include?: GlobPattern;
+
+        /**
+         * A [glob pattern](#GlobPattern) that defines files and folders to exclude. The glob pattern
+         * will be matched against the file paths of resulting matches relative to their workspace. When `undefined`, default excludes will
+         * apply.
+         */
+        exclude?: GlobPattern;
+
+        /**
+         * Whether to use the default and user-configured excludes. Defaults to true.
+         */
+        useDefaultExcludes?: boolean;
+
+        /**
+         * The maximum number of results to search for
+         */
+        maxResults?: number;
+
+        /**
+         * Whether external files that exclude files, like .gitignore, should be respected.
+         * See the vscode setting `"search.useIgnoreFiles"`.
+         */
+        useIgnoreFiles?: boolean;
+
+        /**
+         * Whether global files that exclude files, like .gitignore, should be respected.
+         * See the vscode setting `"search.useGlobalIgnoreFiles"`.
+         */
+        useGlobalIgnoreFiles?: boolean;
+
+        /**
+         * Whether symlinks should be followed while searching.
+         * See the vscode setting `"search.followSymlinks"`.
+         */
+        followSymlinks?: boolean;
+
+        /**
+         * Interpret files using this encoding.
+         * See the vscode setting `"files.encoding"`
+         */
+        encoding?: string;
+
+        /**
+         * Options to specify the size of the result text preview.
+         */
+        previewOptions?: TextSearchPreviewOptions;
+
+        /**
+         * Number of lines of context to include before each match.
+         */
+        beforeContext?: number;
+
+        /**
+         * Number of lines of context to include after each match.
+         */
+        afterContext?: number;
+    }
+
+    /**
+     * A match from a text search
+     */
+    export interface TextSearchMatch {
+        /**
+         * The uri for the matching document.
+         */
+        uri: Uri;
+
+        /**
+         * The range of the match within the document, or multiple ranges for multiple matches.
+         */
+        ranges: Range | Range[];
+
+        /**
+         * A preview of the text match.
+         */
+        preview: TextSearchMatchPreview;
+    }
+
+    /**
+     * A preview of the text result.
+     */
+    export interface TextSearchMatchPreview {
+        /**
+         * The matching lines of text, or a portion of the matching line that contains the match.
+         */
+        text: string;
+
+        /**
+         * The Range within `text` corresponding to the text of the match.
+         * The number of matches must match the TextSearchMatch's range property.
+         */
+        matches: Range | Range[];
+    }
+
+    /**
+     * A line of context surrounding a TextSearchMatch.
+     */
+    export interface TextSearchContext {
+        /**
+         * The uri for the matching document.
+         */
+        uri: Uri;
+
+        /**
+         * One line of text.
+         * previewOptions.charsPerLine applies to this
+         */
+        text: string;
+
+        /**
+         * The line number of this line of context.
+         */
+        lineNumber: number;
+    }
+
+    export type TextSearchResult = TextSearchMatch | TextSearchContext;
+
+    /**
+     * Information collected when text search is complete.
+     */
+    export interface TextSearchComplete {
+        /**
+         * Whether the search hit the limit on the maximum number of search results.
+         * `maxResults` on [`TextSearchOptions`](#TextSearchOptions) specifies the max number of results.
+         * - If exactly that number of matches exist, this should be false.
+         * - If `maxResults` matches are returned and more exist, this should be true.
+         * - If search hits an internal limit which is less than `maxResults`, this should be true.
+         */
+        limitHit?: boolean;
+    }
+
+    export namespace workspace {
+        /**
+         * Find text in files across all [workspace folders] in the workspace
+         * @param query What to search
+         * @param optionsOrCallback
+         * @param callbackOrToken
+         * @param token
+         */
+        export function findTextInFiles(query: TextSearchQuery, optionsOrCallback: FindTextInFilesOptions | ((result: TextSearchResult) => void),
+            callbackOrToken?: CancellationToken | ((result: TextSearchResult) => void), token?: CancellationToken): Promise<TextSearchComplete>;
+    }
+
+}
