+++ packages/plugin/src/theia.proposed.textSearchProvider.d.ts
@@ -1,0 +1,145 @@
+// *****************************************************************************
+// Copyright (C) 2023 Ericsson and others.
+//
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License v. 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0.
+//
+// This Source Code may also be made available under the following Secondary
+// Licenses when the conditions for such availability set forth in the Eclipse
+// Public License v. 2.0 are satisfied: GNU General Public License, version 2
+// with the GNU Classpath Exception which is available at
+// https://www.gnu.org/software/classpath/license.html.
+//
+// SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0
+// *****************************************************************************
+
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+// code copied and modified from https://github.com/microsoft/vscode/blob/1.77.0/src/vscode-dts/vscode.proposed.textSearchQuery.d.ts
+
+export module '@theia/plugin' {
+
+    /**
+     * The parameters of a query for text search.
+     */
+    export interface TextSearchQuery {
+        /**
+         * The text pattern to search for.
+         */
+        pattern: string;
+
+        /**
+         * Whether or not `pattern` should match multiple lines of text.
+         */
+        isMultiline?: boolean;
+
+        /**
+         * Whether or not `pattern` should be interpreted as a regular expression.
+         */
+        isRegExp?: boolean;
+
+        /**
+         * Whether or not the search should be case-sensitive.
+         */
+        isCaseSensitive?: boolean;
+
+        /**
+         * Whether or not to search for whole word matches only.
+         */
+        isWordMatch?: boolean;
+    }
+
+    /**
+     * A match from a text search
+     */
+    export interface TextSearchMatch {
+        /**
+         * The uri for the matching document.
+         */
+        uri: Uri;
+
+        /**
+         * The range of the match within the document, or multiple ranges for multiple matches.
+         */
+        ranges: Range | Range[];
+
+        /**
+         * A preview of the text match.
+         */
+        preview: TextSearchMatchPreview;
+    }
+
+    /**
+     * A preview of the text result.
+     */
+    export interface TextSearchMatchPreview {
+        /**
+         * The matching lines of text, or a portion of the matching line that contains the match.
+         */
+        text: string;
+
+        /**
+         * The Range within `text` corresponding to the text of the match.
+         * The number of matches must match the TextSearchMatch's range property.
+         */
+        matches: Range | Range[];
+    }
+
+    /**
+     * Options to specify the size of the result text preview.
+     * These options don't affect the size of the match itself, just the amount of preview text.
+     */
+    export interface TextSearchPreviewOptions {
+        /**
+         * The maximum number of lines in the preview.
+         * Only search providers that support multiline search will ever return more than one line in the match.
+         */
+        matchLines: number;
+
+        /**
+         * The maximum number of characters included per line.
+         */
+        charsPerLine: number;
+    }
+
+    /**
+     * A line of context surrounding a TextSearchMatch.
+     */
+    export interface TextSearchContext {
+        /**
+         * The uri for the matching document.
+         */
+        uri: Uri;
+
+        /**
+         * One line of text.
+         * previewOptions.charsPerLine applies to this
+         */
+        text: string;
+
+        /**
+         * The line number of this line of context.
+         */
+        lineNumber: number;
+    }
+
+    export type TextSearchResult = TextSearchMatch | TextSearchContext;
+
+    /**
+     * Information collected when text search is complete.
+     */
+    export interface TextSearchComplete {
+        /**
+         * Whether the search hit the limit on the maximum number of search results.
+         * `maxResults` on [`TextSearchOptions`](#TextSearchOptions) specifies the max number of results.
+         * - If exactly that number of matches exist, this should be false.
+         * - If `maxResults` matches are returned and more exist, this should be true.
+         * - If search hits an internal limit which is less than `maxResults`, this should be true.
+         */
+        limitHit?: boolean;
+    }
+
+}
